chrome.permissions.request("tabs")
This delays asking for tab.url access until
the user clicks on the icon. Once permission
is granted, it sticks: background.js will be
able to get the number of comments from
comntr.live. Context menus don't need this
permission.

Tested on Chrome and Firefox. Not tested on
Firefox Android.